+++ src/components/Bars/Bars.tsx
@@ -79,9 +79,9 @@ interface IBarsProps extends StandardProps, React.SVGProps<SVGGElement> {
 	 */
 	//xScale: (x: string) => number | undefined;
 	xScale:
-		| d3Scale.ScaleBand<string>
-		| d3Scale.ScaleBand<number>
-		| d3Scale.ScalePoint<string>;
+	| d3Scale.ScaleBand<string>
+	| d3Scale.ScaleBand<number>
+	| d3Scale.ScalePoint<string>;
 
 	/** The field we should look up your x data by.*/
 	xField: string;
@@ -94,9 +94,9 @@ interface IBarsProps extends StandardProps, React.SVGProps<SVGGElement> {
 	 * Lucid exposes the \`lucid.d3Scale\` library for use here.
 	 * */
 	yScale:
-		| d3Scale.ScaleBand<string | number>
-		| d3Scale.ScalePoint<string | number>
-		| d3Scale.ScaleLinear<number, number>;
+	| d3Scale.ScaleBand<string | number>
+	| d3Scale.ScalePoint<string | number>
+	| d3Scale.ScaleLinear<number, number>;
 
 	/** The field(s) we should look up your y data by.
 	 * Each entry represents a series.
@@ -408,11 +408,11 @@ export class Bars extends PureComponent<IBarsProps, IBarsState> {
 									isStacked
 										? xScale(data[seriesIndex][xField] as any)
 										: // prettier-ignore
-										  //@ts-ignore
-										  innerXScale(pointsIndex) +
+										//@ts-ignore
+										innerXScale(pointsIndex) +
 										// prettier-ignore
-										  //@ts-ignore
-										  xScale(data[seriesIndex][xField] as any)
+										//@ts-ignore
+										xScale(data[seriesIndex][xField] as any)
 								}
 								y={yScale(end)}
 								height={(yScale(start) as any) - (yScale(end) as any)}
@@ -430,11 +430,11 @@ export class Bars extends PureComponent<IBarsProps, IBarsState> {
 							height={
 								isStacked
 									? // prettier-ignore
-									  //@ts-ignore
-									  yScale.range()[0] - yScale(_.last(series)[1])
+									//@ts-ignore
+									yScale.range()[0] - yScale(_.last(series)[1])
 									: // prettier-ignore
-									  //@ts-ignore
-									  yScale.range()[0] - yScale(_.max(_.flatten(series)))
+									//@ts-ignore
+									yScale.range()[0] - yScale(_.max(_.flatten(series)))
 							}
 							width={xScale.bandwidth()}
 							x={xScale(data[seriesIndex][xField] as any)}
@@ -474,8 +474,8 @@ export interface IPureToolTipsProps extends StandardProps {
 	//onMouseEnter: (event: MouseEvent<SVGRectElement, MouseEvent>) => void
 
 	onMouseOut?:
-		| ((event: React.MouseEvent<SVGRectElement, MouseEvent>) => void)
-		| undefined;
+	| ((event: React.MouseEvent<SVGRectElement, MouseEvent>) => void)
+	| undefined;
 
 	xFormatter: (d: Date, seriesIndex: number) => string;
